refactor(f39): modularize replay_harness.py into package
Split monolithic scripts/f39/replay_harness.py (~950 lines) into a
proper package with 5 focused modules: models, prefilter, single_replay,
chain_replay, cli. The old file is now a 10-line stub that delegates to
replay_harness.cli.main. Backwards-compatible re-exports preserved in
__init__.py.

Exception handling narrowed per OI-1437: OSError in os.unlink finally
blocks and (OSError, json.JSONDecodeError) in chain-type detection are
caught and logged at debug level rather than silently swallowed or
propagated. 4/4 regression tests pass.

Dispatch-ID: dispatch-fwd-01
